Fix bug for x64 cpus for reg allocation
Bug was that 255 is less than 64*sizeof(void*).
This could lead to overflow of Uint8 and
some allocated registers would clash with invaild reg value.

+++ src/Engine/ScriptBind.h
@@ -148,7 +148,7 @@ struct ParserWriter
 	ReservedCrossRefrenece<ScriptText, ScriptRef> refTexts;
 
 	/// index of used script registers.
-	Uint8 regIndexUsed;
+	size_t regIndexUsed;
 	/// negative index of used const values.
 	int constIndexUsed;
 
@@ -157,7 +157,7 @@ struct ParserWriter
 
 	/// Constructor.
 	ParserWriter(
-			Uint8 regUsed,
+			size_t regUsed,
 			ScriptContainerBase& c,
 			const ScriptParserBase& d);
